Added Edge functionality
Edges can now be added to graphs. It's a little inadequete, considering clicking a source node, then the destination node is necessary. A drag from one node to the next would be more ergonomic, however I don't know how to use that functionality. Also the mouse events are being handled by the QGraphicView, which may be suboptimal. The events may need to be handled by the Nodes and Edges themselves. It might make it a bit cleaner, considering a search has to be done in the QGraphicView for items under the mouse. Overall the usability of the program is subpar, however the functionality is there.

Reworking the functionality to be more human friendly and starting on the viewing components for the algorithms (Stack, Queue, Table) are next.